+++ test/hipcub/test_utils_assertions.hpp
@@ -238,6 +238,82 @@ inline void assert_bit_eq(const std::vector<T>& result, const std::vector<T>& ex
     }
 }
 
+#if HIPCUB_IS_INT128_ENABLED
+inline void assert_bit_eq(const std::vector<__int128_t>& result,
+                          const std::vector<__int128_t>& expected)
+{
+    ASSERT_EQ(result.size(), expected.size());
+
+    auto to_string = [](__int128_t value)
+    {
+        static const char* charmap = "0123456789";
+
+        std::string result;
+        result.reserve(41); // max. 40 digits possible ( uint64_t has 20) plus sign
+        __uint128_t helper = (value < 0) ? -value : value;
+
+        do
+        {
+            result += charmap[helper % 10];
+            helper /= 10;
+        }
+        while(helper);
+        if(value < 0)
+        {
+            result += "-";
+        }
+        std::reverse(result.begin(), result.end());
+        return result;
+    };
+
+    for(size_t i = 0; i < result.size(); i++)
+    {
+        if(!bit_equal(result[i], expected[i]))
+        {
+            FAIL() << "Expected strict/bitwise equality of these values: " << std::endl
+                   << "     result[i]: " << to_string(result[i]) << std::endl
+                   << "     expected[i]: " << to_string(expected[i]) << std::endl
+                   << "where index = " << i;
+        }
+    }
+}
+
+inline void assert_bit_eq(const std::vector<__uint128_t>& result,
+                          const std::vector<__uint128_t>& expected)
+{
+    ASSERT_EQ(result.size(), expected.size());
+
+    auto to_string = [](__uint128_t value)
+    {
+        static const char* charmap = "0123456789";
+
+        std::string result;
+        result.reserve(40); // max. 40 digits possible ( uint64_t has 20)
+        __uint128_t helper = value;
+
+        do
+        {
+            result += charmap[helper % 10];
+            helper /= 10;
+        }
+        while(helper);
+        std::reverse(result.begin(), result.end());
+        return result;
+    };
+
+    for(size_t i = 0; i < result.size(); i++)
+    {
+        if(!bit_equal(result[i], expected[i]))
+        {
+            FAIL() << "Expected strict/bitwise equality of these values: " << std::endl
+                   << "     result[i]: " << to_string(result[i]) << std::endl
+                   << "     expected[i]: " << to_string(expected[i]) << std::endl
+                   << "where index = " << i;
+        }
+    }
+}
+#endif //HIPCUB_IS_INT128_ENABLED
+
 /// Compile-time assertion for type equality of two objects.
 template<class ExpectedT, class ActualT>
 inline void assert_type(ExpectedT /*obj1*/, ActualT /*obj2*/)